# openapi-generated-aspnetcore/src/Org.OpenAPITools/Controllers/UsersApi.cs
/*
 * PIM-SWKS SoSe2024 REST API
 *
 * PIM-SWKS SoSe2024 REST API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using Microsoft.AspNetCore.Http;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using Org.OpenAPITools.Attributes;
using Org.OpenAPITools.Models;

namespace Org.OpenAPITools.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class UsersApiController : ControllerBase
    { 
        /// <summary>
        /// Create a new user
        /// </summary>
        /// <param name="user"></param>
        /// <response code="201">Created</response>
        /// <response code="401">Unauthorized</response>
        /// <response code="403">Forbidden</response>
        [HttpPost]
        [Route("/users")]
        [Authorize(Policy = "apikeyAuth")]
        [Consumes("application/json")]
        [ValidateModelState]
        [SwaggerOperation("CreateUser")]
        [SwaggerResponse(statusCode: 201, type: typeof(User), description: "Created")]
        public virtual IActionResult CreateUser([FromBody]User user)
        {

            //TODO: Uncomment the next line to return response 201 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(201, default(User));
            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401);
            //TODO: Uncomment the next line to return response 403 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(403);
            string exampleJson = null;
            exampleJson = "{\n  \"id\" : 1,\n  \"email\" : \"email@example.com\",\n  \"username\" : \"admin\"\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<User>(exampleJson)
            : default(User);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Delete a user by ID
        /// </summary>
        /// <param name="id">ID of the user to delete</param>
        /// <response code="200">OK</response>
        /// <response code="401">Unauthorized</response>
        /// <response code="403">Forbidden</response>
        /// <response code="404">If the user does not exist</response>
        [HttpDelete]
        [Route("/users/{id}")]
        [Authorize(Policy = "apikeyAuth")]
        [ValidateModelState]
        [SwaggerOperation("DeleteUserById")]
        [SwaggerResponse(statusCode: 200, type: typeof(User), description: "OK")]
        public virtual IActionResult DeleteUserById([FromRoute (Name = "id")][Required]long id)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(User));
            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401);
            //TODO: Uncomment the next line to return response 403 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(403);
            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404);
            string exampleJson = null;
            exampleJson = "{\n  \"id\" : 1,\n  \"email\" : \"email@example.com\",\n  \"username\" : \"admin\"\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<User>(exampleJson)
            : default(User);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Returns all comments authored by this user
        /// </summary>
        /// <param name="id">ID of the user to return comments for</param>
        /// <response code="200">OK</response>
        /// <response code="404">If the user does not exist</response>
        [HttpGet]
        [Route("/users/{id}/comments")]
        [ValidateModelState]
        [SwaggerOperation("GetCommentsForUser")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<Comment>), description: "OK")]
        public virtual IActionResult GetCommentsForUser([FromRoute (Name = "id")][Required]long id)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(List<Comment>));
            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404);
            string exampleJson = null;
            exampleJson = "{\n  \"id\" : 1,\n  \"postId\" : 1,\n  \"authorId\" : 1,\n  \"content\" : \"This is a comment\"\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<List<Comment>>(exampleJson)
            : default(List<Comment>);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Returns all posts authored by this user
        /// </summary>
        /// <param name="id">ID of the user to return posts for</param>
        /// <response code="200">OK</response>
        /// <response code="404">If the user does not exist</response>
        [HttpGet]
        [Route("/users/{id}/posts")]
        [ValidateModelState]
        [SwaggerOperation("GetPostsForUser")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<Post>), description: "OK")]
        public virtual IActionResult GetPostsForUser([FromRoute (Name = "id")][Required]long id)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(List<Post>));
            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404);
            string exampleJson = null;
            exampleJson = "{\n  \"id\" : 1,\n  \"title\" : \"Hello World\",\n  \"authorId\" : 1,\n  \"content\" : \"This is a post\"\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<List<Post>>(exampleJson)
            : default(List<Post>);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Returns a user by ID
        /// </summary>
        /// <param name="id">ID of the user to return</param>
        /// <response code="200">OK</response>
        /// <response code="404">If the user does not exist</response>
        [HttpGet]
        [Route("/users/{id}")]
        [ValidateModelState]
        [SwaggerOperation("GetUserById")]
        [SwaggerResponse(statusCode: 200, type: typeof(User), description: "OK")]
        public virtual IActionResult GetUserById([FromRoute (Name = "id")][Required]long id)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(User));
            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404);
            string exampleJson = null;
            exampleJson = "{\n  \"id\" : 1,\n  \"email\" : \"email@example.com\",\n  \"username\" : \"admin\"\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<User>(exampleJson)
            : default(User);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Returns all users
        /// </summary>
        /// <response code="200">OK</response>
        [HttpGet]
        [Route("/users")]
        [ValidateModelState]
        [SwaggerOperation("GetUsers")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<User>), description: "OK")]
        public virtual IActionResult GetUsers()
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(List<User>));
            string exampleJson = null;
            exampleJson = "{\n  \"id\" : 1,\n  \"email\" : \"email@example.com\",\n  \"username\" : \"admin\"\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<List<User>>(exampleJson)
            : default(List<User>);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Update a user by ID
        /// </summary>
        /// <param name="id">ID of the user to update</param>
        /// <param name="user"></param>
        /// <response code="200">OK</response>
        /// <response code="401">Unauthorized</response>
        /// <response code="403">Forbidden</response>
        /// <response code="404">If the user does not exist</response>
        [HttpPut]
        [Route("/users/{id}")]
        [Authorize(Policy = "apikeyAuth")]
        [Consumes("application/json")]
        [ValidateModelState]
        [SwaggerOperation("UpdateUserById")]
        [SwaggerResponse(statusCode: 200, type: typeof(User), description: "OK")]
        public virtual IActionResult UpdateUserById([FromRoute (Name = "id")][Required]long id, [FromBody]User user)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(User));
            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401);
            //TODO: Uncomment the next line to return response 403 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(403);
            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404);
            string exampleJson = null;
            exampleJson = "{\n  \"id\" : 1,\n  \"email\" : \"email@example.com\",\n  \"username\" : \"admin\"\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<User>(exampleJson)
            : default(User);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }
    }
}