# openapi-generated-aspnetcore/src/Org.OpenAPITools/wwwroot/openapi-original.json
{
  "openapi" : "3.0.3",
  "info" : {
    "description" : "PIM-SWKS SoSe2024 REST API",
    "title" : "PIM-SWKS SoSe2024 REST API",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "description" : "Local development server",
    "url" : "http://localhost:{Port}",
    "variables" : {
      "Port" : {
        "default" : "18080",
        "description" : "The port number (18080 - rest-jakarta, 28080 - rest-springboot, 38080 - rest-aspnet)",
        "enum" : [ "18080", "28080", "38080", "8080" ]
      }
    }
  }, {
    "description" : "Production server",
    "url" : "https://pim-swks.nekanu.com"
  } ],
...
  "paths" : {
...
    "/users" : {
      "get" : {
        "operationId" : "getUsers",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/User"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Returns all users",
        "tags" : [ "users" ]
      },
      "post" : {
        "operationId" : "createUser",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/User"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              }
            },
            "description" : "Created"
          },
          "401" : {
            "description" : "Unauthorized",
            "headers" : {
              "WWW-Authenticate" : {
                "description" : "The authentication method that should be used",
                "explode" : false,
                "schema" : {
                  "type" : "string"
                },
                "style" : "simple"
              }
            }
          },
          "403" : {
            "description" : "Forbidden"
          }
        },
        "security" : [ {
          "apikeyAuth" : [ ]
        } ],
        "summary" : "Create a new user",
        "tags" : [ "users" ]
      }
    },
    "/users/{id}" : {
      "delete" : {
        "operationId" : "deleteUserById",
        "parameters" : [ {
          "description" : "ID of the user to delete",
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "description" : "Unauthorized",
            "headers" : {
              "WWW-Authenticate" : {
                "description" : "The authentication method that should be used",
                "explode" : false,
                "schema" : {
                  "type" : "string"
                },
                "style" : "simple"
              }
            }
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "If the user does not exist"
          }
        },
        "security" : [ {
          "apikeyAuth" : [ ]
        } ],
        "summary" : "Delete a user by ID",
        "tags" : [ "users" ]
      },
      "get" : {
        "operationId" : "getUserById",
        "parameters" : [ {
          "description" : "ID of the user to return",
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              }
            },
            "description" : "OK"
          },
          "404" : {
            "description" : "If the user does not exist"
          }
        },
        "summary" : "Returns a user by ID",
        "tags" : [ "users" ]
      },
      "put" : {
        "operationId" : "updateUserById",
        "parameters" : [ {
          "description" : "ID of the user to update",
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          },
          "style" : "simple"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/User"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "description" : "Unauthorized",
            "headers" : {
              "WWW-Authenticate" : {
                "description" : "The authentication method that should be used",
                "explode" : false,
                "schema" : {
                  "type" : "string"
                },
                "style" : "simple"
              }
            }
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "If the user does not exist"
          }
        },
        "security" : [ {
          "apikeyAuth" : [ ]
        } ],
        "summary" : "Update a user by ID",
        "tags" : [ "users" ]
      }
    },
...
    "/users/{id}/comments" : {
      "get" : {
        "operationId" : "getCommentsForUser",
        "parameters" : [ {
          "description" : "ID of the user to return comments for",
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/Comment"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          },
          "404" : {
            "description" : "If the user does not exist"
          }
        },
        "summary" : "Returns all comments authored by this user",
        "tags" : [ "comments", "users" ]
      }
    },
    "/users/{id}/posts" : {
      "get" : {
        "operationId" : "getPostsForUser",
        "parameters" : [ {
          "description" : "ID of the user to return posts for",
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/Post"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          },
          "404" : {
            "description" : "If the user does not exist"
          }
        },
        "summary" : "Returns all posts authored by this user",
        "tags" : [ "posts", "users" ]
      }
    }
  },
  "components" : {
...
    "schemas" : {
      "Post" : {
        "example" : {
          "id" : 1,
          "title" : "Hello World",
          "authorId" : 1,
          "content" : "This is a post"
        },
        "properties" : {
          "id" : {
            "example" : 1,
            "format" : "uint64",
            "readOnly" : true,
            "type" : "integer"
          },
          "title" : {
            "example" : "Hello World",
            "maxLength" : 512,
            "minLength" : 3,
            "type" : "string"
          },
          "content" : {
            "example" : "This is a post",
            "format" : "markdown",
            "type" : "string"
          },
          "authorId" : {
            "description" : "ID of the author of the post",
            "example" : 1,
            "format" : "uint64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "User" : {
        "example" : {
          "id" : 1,
          "email" : "email@example.com",
          "username" : "admin"
        },
        "properties" : {
          "id" : {
            "example" : 1,
            "format" : "uint64",
            "minimum" : 1,
            "readOnly" : true,
            "type" : "integer"
          },
          "username" : {
            "example" : "admin",
            "maxLength" : 64,
            "minLength" : 3,
            "pattern" : "^[a-zA-Z0-9-_.]{3,64}$",
            "type" : "string"
          },
          "email" : {
            "example" : "email@example.com",
            "format" : "email",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "Comment" : {
        "example" : {
          "id" : 1,
          "postId" : 1,
          "authorId" : 1,
          "content" : "This is a comment"
        },
        "properties" : {
          "id" : {
            "example" : 1,
            "format" : "uint64",
            "readOnly" : true,
            "type" : "integer"
          },
          "content" : {
            "example" : "This is a comment",
            "type" : "string"
          },
          "authorId" : {
            "description" : "ID of the author of the comment",
            "example" : 1,
            "format" : "uint64",
            "type" : "integer"
          },
          "postId" : {
            "description" : "ID of the post the comment is for",
            "example" : 1,
            "format" : "uint64",
            "type" : "integer"
          }
        },
        "type" : "object"
      }
    },
    "securitySchemes" : {
      "apikeyAuth" : {
        "in" : "header",
        "name" : "X-API-KEY",
        "type" : "apiKey"
      }
    }
  }
}